feat(app/v3): add node

+++ app/model/task.go
@@ -5,7 +5,7 @@ type Task struct {
 
 	Name   string `json:"name" form:"name"`
 	TeamID uint   `json:"-"`
-	NodeID int64  `json:"node_id" form:"node_id"`
+	NodeID uint   `json:"node_id" form:"node_id"`
 	Job    *Job   `json:"job,omitempty"`
 	Files  JSON   `json:"files"`
 	Extra  JSON   `json:"extra"`
